bot/requirements.txt
python-telegram-bot==21.6
redis==5.0.8
python-dotenv==1.0.1
openai==1.51.0
